--- addons/bombs/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\ace\addons\bombs
