--- addons/bombs/CfgEventhandlers.hpp
@@ -1,24 +1,0 @@
-
-class Extended_PreStart_EventHandlers {
-    class ADDON {
-        init = QUOTE(call COMPILE_FILE(XEH_preStart));
-    };
-};
-
-class Extended_PreInit_EventHandlers {
-    class ADDON    {
-        init = QUOTE(call COMPILE_FILE(XEH_preInit));
-    };
-};
-
-class Extended_PostInit_EventHandlers {
-    class ADDON {
-        init = QUOTE(call COMPILE_FILE(XEH_postInit));
-    };
-};
-
-class Extended_FiredBIS_EventHandlers {
-    class All {
-        ADDON = QUOTE(_this call FUNC(onFired));
-    };
-};
