--- addons/bombs/XEH_PREP.hpp
@@ -1,7 +1,0 @@
-PREP(onFired);
-PREP(guidancePFH);
-PREP(steering);
-PREP(adjustTarget);
-
-PREP(seeker_GPS);
-PREP(seeker_SALH);
