--- addons/bombs/XEH_postInit.sqf
@@ -1,1 +1,0 @@
-#include "script_component.hpp"
