--- addons/bombs/XEH_preStart.sqf
@@ -1,3 +1,0 @@
-#include "script_component.hpp"
-
-#include "XEH_PREP.hpp"
