--- addons/bombs/config.cpp
@@ -1,28 +1,0 @@
-#include "script_component.hpp"
-
-class CfgPatches {
-    class ADDON {
-        name = COMPONENT_NAME;
-        units[] = {};
-        weapons[] = {};
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"ace_missileguidance"};
-        author = ECSTRING(common,ACETeam);
-        authors[] = {"xrufix"};
-        url = ECSTRING(main,URL);
-        VERSION_CONFIG;
-    };
-};
-
-#include "CfgEventhandlers.hpp"
-#include "CfgAmmo.hpp"
-#include "CfgWeapons.hpp"
-
-class GVAR(seekerTypes) {
-    class SALH {
-        function = QFUNC(seeker_SALH);
-    };
-    class GPS {
-        function = QFUNC(seeker_GPS);
-    };
-};
