+++ addons/missileguidance/CfgAmmo.hpp
@@ -1,4 +1,5 @@
 class CfgAmmo {
+    class RocketPods;
     class MissileBase;
 
     class M_PG_AT: MissileBase {};
@@ -102,5 +103,12 @@ class CfgAmmo {
 
         //Explicity add guidance config
         class ADDON: ADDON {};
+    };
+
+    class Bomb_04_Plane_CAS_01_F: RocketPods {
+        ace_laser_canSelect = 1;
+        ace_laser_showHud = 1;
+        canLock = 0;
+        weaponLockSystem = 0;
     };
 };
