+++ addons/missileguidance/functions/fnc_changeMissileDirection.sqf
@@ -22,5 +22,9 @@ params ["_projectile", "_v"];
 private _l = sqrt ((_v select 0) ^ 2 + (_v select 1) ^ 2);
 private _r = -(_v select 2) / _l;
 
-_projectile setVectorDirAndUp [ _v, [(_v select 0) * _r,(_v select 1) * _r, _l] ];
-_projectile setVelocity (_v vectorMultiply (vectorMagnitude (velocity _projectile)));
+_projectile setVectorDirAndUp [ _v, [(_v select 0) * _r, (_v select 1) * _r, _l] ];
+
+// change Velocity only for missiles, let engine handle bombs.
+if (!_isBomb) then {
+	_projectile setVelocity (_v vectorMultiply (vectorMagnitude (velocity _projectile)));
+};
