+++ addons/missileguidance/script_component.hpp
@@ -2,9 +2,9 @@
 #define COMPONENT_BEAUTIFIED Missile Guidance
 #include "\z\ace\addons\main\script_mod.hpp"
 
-// #define DRAW_GUIDANCE_INFO
-// #define DEBUG_MODE_FULL
-// #define DISABLE_COMPILE_CACHE
+#define DRAW_GUIDANCE_INFO
+#define DEBUG_MODE_FULL
+#define DISABLE_COMPILE_CACHE
 // #define ENABLE_PERFORMANCE_COUNTERS
 
 #ifdef DEBUG_ENABLED_MISSILEGUIDANCE
